Clean up elibc headers and libc usage
