+++ include/openenclave/elibc/pthread.h
@@ -140,7 +140,16 @@ void* oe_pthread_getspecific(oe_pthread_key_t key);
 
 #if defined(OE_NEED_STDC_NAMES)
 
-#include "bits/pthread.h"
+#include "bits/pthread_def.h"
+
+#include "bits/pthread_cond.h"
+#include "bits/pthread_create.h"
+#include "bits/pthread_equal.h"
+#include "bits/pthread_key.h"
+#include "bits/pthread_mutex.h"
+#include "bits/pthread_once.h"
+#include "bits/pthread_rwlock.h"
+#include "bits/pthread_spin.h"
 
 #endif /* defined(OE_NEED_STDC_NAMES) */
