+++ libc/malloc.c
@@ -12,7 +12,6 @@
 #undef OE_INLINE
 #define OE_INLINE
 #endif
-#define OE_USE_MUSL_DEFS
 #include <openenclave/elibc/bits/malloc.h>
 #if defined(__UNDEF_OE_NEED_STDC_NAMES)
 #undef OE_NEED_STDC_NAMES
