+++ libc/pthread.c
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+#include <openenclave/elibc/pthread.h>
 #include <openenclave/elibc/string.h>
 #include <openenclave/enclave.h>
 #include <openenclave/internal/defs.h>
@@ -21,8 +22,13 @@
 #undef OE_INLINE
 #define OE_INLINE
 #endif
-#define OE_USE_MUSL_DEFS
-#include <openenclave/elibc/pthread.h>
+#include <openenclave/elibc/bits/pthread_cond.h>
+#include <openenclave/elibc/bits/pthread_equal.h>
+#include <openenclave/elibc/bits/pthread_key.h>
+#include <openenclave/elibc/bits/pthread_mutex.h>
+#include <openenclave/elibc/bits/pthread_once.h>
+#include <openenclave/elibc/bits/pthread_rwlock.h>
+#include <openenclave/elibc/bits/pthread_spin.h>
 #if defined(__UNDEF_OE_NEED_STDC_NAMES)
 #undef OE_NEED_STDC_NAMES
 #undef __UNDEF_OE_NEED_STDC_NAMES
